Manual string constuction from range
